pyproject.toml
[project]
name = "gemini-multimodal-audio-upload"
version = "0.1.0"
description = "Upload audio to gemini for analysis and feedback.  Takes audio, json context and prompt."
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "google-generativeai>=0.8.5",
    "mcp>=1.23.1",
    "python-dotenv>=1.2.1",
]

[dependency-groups]
dev = [
    "pre-commit>=4.5.0",
    "pylint>=4.0.4",
]
